Add delinting for MSVC

+++ ezgz_test.cpp
@@ -18,7 +18,7 @@ template <int Size>
 struct InputHelper : EzGz::Detail::ByteInput<SettingsWithInputSize<Size>> {
 	InputHelper(std::span<const uint8_t> source)
 	: EzGz::Detail::ByteInput<SettingsWithInputSize<Size>>([source, position = 0] (std::span<uint8_t> toFill) mutable -> int {
-		int filling = std::min(source.size() - position, toFill.size());
+		int filling = static_cast<int>(std::min(source.size() - position, toFill.size()));
 //		std::cout << "Providing " << filling << " bytes of data, " << (source.size() - position - filling) << " left" << std::endl;
 		if(filling != 0)
 			memcpy(toFill.data(), &source[position], filling);
@@ -64,7 +64,7 @@ int main(int, char**) {
 		}
 
 		{
-			unsigned int twoBytes = byteReader.getBytes(2);
+			unsigned int twoBytes = static_cast<unsigned int>(byteReader.getBytes(2));
 			doATest(twoBytes, 0b1010101010101010u);
 		}
 
@@ -94,7 +94,7 @@ int main(int, char**) {
 			while (read < tryingToRead) {
 				auto readTenBytes = reader.getRange(tryingToRead - read);
 				doATest(readTenBytes.size() < tryingToRead, true);
-				read += readTenBytes.size();
+				read += static_cast<int>(readTenBytes.size());
 			}
 			doATest(read, tryingToRead);
 		}
@@ -231,8 +231,8 @@ int main(int, char**) {
 		auto inspectStart = [&doATest, shouldBe, position = 0] (std::span<const char> reading) mutable -> int {
 			std::string_view correctPart = shouldBe.substr(position, reading.size());
 			doATest(std::string_view(reading.data(), reading.size()), correctPart);
-			position += reading.size();
-			return reading.size();
+			position += static_cast<int>(reading.size());
+			return static_cast<int>(reading.size());
 		};
 
 		{
